Ignore sqlite

+++ src/sentry/south_migrations/0416_auto__del_field_identityprovider_organization__add_field_identityprovi.py
@@ -3,6 +3,7 @@ from south.utils import datetime_utils as datetime
 from south.db import db
 from south.v2 import SchemaMigration
 from django.db import models
+from sentry.utils.db import is_sqlite
 
 
 class Migration(SchemaMigration):
@@ -12,6 +13,8 @@ class Migration(SchemaMigration):
     is_dangerous = False
 
     def forwards(self, orm):
+        if is_sqlite():
+            return
         db.delete_foreign_key(u'sentry_identityprovider', 'organization_id')
         db.alter_column(
             u'sentry_identityprovider',
